Change flag size to u_int32_t
git-svn-id: https://www.dnssec-tools.org/svn/dnssec-tools/trunk@4009 c4029f9c-b713-0410-a448-fae9e31a314a

+++ dnssec-tools/validator/libval/val_assertion.h
@@ -38,7 +38,7 @@ int             add_to_qfq_chain(val_context_t *context,
                                  struct queries_for_query **queries,
                                  u_char * name_n, const u_int16_t type_h,
                                  const u_int16_t class_h, 
-                                 const u_int8_t flags,
+                                 const u_int32_t flags,
                                  struct queries_for_query **added_qfq);
 void            free_authentication_chain(struct val_digested_auth_chain
                                           *assertions);
@@ -48,14 +48,14 @@ int             val_istrusted(val_status_t val_status);
 int             val_isvalidated(val_status_t val_status);
 int             val_does_not_exist(val_status_t status); 
 int             get_zse(val_context_t * ctx, u_int8_t * name_n, 
-                        u_int8_t flags, u_int16_t *status, u_int8_t ** match_ptr, u_int32_t *ttl_x);
+                        u_int32_t flags, u_int16_t *status, u_int8_t ** match_ptr, u_int32_t *ttl_x);
 int             find_trust_point(val_context_t * ctx, u_int8_t * zone_n, 
                                  u_int8_t ** matched_zone, u_int32_t *ttl_x);
 void            val_free_result_chain(struct val_result_chain *results);
 #ifdef LIBVAL_DLV
 int             check_anc_proof(val_context_t *context,
                                 struct val_query_chain *q, 
-                                u_int8_t flags,
+                                u_int32_t flags,
                                 u_int8_t *name_n, 
                                 int check_wildcard,
                                 int *matches);
@@ -64,7 +64,7 @@ int             try_chase_query(val_context_t * context,
                                 u_char * domain_name_n,
                                 const u_int16_t q_class,
                                 const u_int16_t type,
-                                const u_int8_t flags,
+                                const u_int32_t flags,
                                 struct queries_for_query **queries,
                                 struct val_result_chain **results,
                                 int *done);
@@ -72,7 +72,7 @@ int             val_resolve_and_check(val_context_t * context,
                                       u_char * domain_name,
                                       const u_int16_t class,
                                       const u_int16_t type,
-                                      const u_int8_t flags,
+                                      const u_int32_t flags,
                                       struct val_result_chain **results);
 
 #endif
